+++ fbw-a380x/src/systems/instruments/src/OITlegacy/OitLegacy.tsx
@@ -103,6 +103,7 @@ export const OitEfbWrapper: React.FC<OitEfbWrapperProps> = ({ eventBus }) => {
             satcom: false,
             latLonExtend: false,
             rmpVhfSpacing: false,
+            avionicsTestMode: true,
           },
           realism: {
             mcduKeyboard: false,